connectionReleaseDelay PoolOption
Close idle connections that wait in the pool for connectionReleaseDelay
milliseconds. Defaults to 0 = keep them forever.

The option name is taken from vertx-mysql-postgresql-client:
https://github.com/vert-x3/vertx-mysql-postgresql-client/blob/3.9.1/vertx-mysql-postgresql-client-jasync/src/main/java/io/vertx/ext/asyncsql/impl/pool/AsyncC$

Implementation
If connectionReleaseDelay is > 0:
When a connection is put into the pool create a timer
that closes the connection and removes it from the pool: expire().
If a connection in the pool is acquired before the timer
fires the timer is cancelled: cancelIdleTimer().

In PoolOptions
* add connectionReleaseDelay option
* fix and unit test equals() and hashCode()

In ConnectionPool
* add constructor that takes PoolOptions to avoid long parameter lists
* add package-private allSize() for unit testing
* rename and move release(PooledConnection) to PooledConnection.addToPool()
  to avoid a name that is misleading if connectionReleaseDelay exists

Signed-off-by: Julian Ladisch <eclipse.org-rtn@ladisch.de>