Fix hostname when connecting to proxy and postgres
PgConnectionAutoRetryTest connects to the proxy using the
hostname returned by ContainerPgRule.
This differs from `localhost` in some CI environments, see #854 and
https://www.testcontainers.org/features/networking/#getting-the-container-host
This needs to be changed to `localhost`, the proxy hostname,
to make the tests succeed.

ConnectionAutoRetryTestBase connects to the database using
`localhost` as hostname. The hostname of the database differs
from `localhost` in some CI environments.
Use the hostname returned by ContainerPgRule to make the tests succeed.

Fixes #854

Signed-off-by: Julian Ladisch <julianladisch@users.noreply.github.com>