When a user gets a connection frompgPool some milliseconds behind IDLE expiration time, then "sometimes" when you are going to use the connection, you get the following error from the server:
Fail to read any response from the server, the underlying connection might get lost unexpectedly.

So, looks that the connection it's available when you request it, but then disappears when you use it.

Expected behavior
The connection should be validated (maybe by a query), before return an available connection.

Actual behavior
An end-user could get a connection that is not available anymore.

Signed-off-by: Pablo Gonzalez granados <pagonzal@redhat.com>