Add NullValue class
Update Tuple so that invoking addXXXX with null switches to adding NullValue

NullValue is a public VertxGen type so that it's convenient for users.
For example: Tuple.of("foo", 5, NullValue.JsonObject)

But when Tuple.getValue is invoked and the storage has a NullValue, we return null instead.

This allows to keep the code compatible with previous versions.

Signed-off-by: Thomas Segismont <tsegismont@gmail.com>