Prototype support of encrypted login
During SSL handshake, outbound and inbound messages are encapsulated in TDS packets.
So we cannot use the Vert.x built-in SSL upgrade for NetClient.
Instead, we must put handler before the SSLHandler that will intercept encrypted traffic.

After the handshake, the interceptor must be removed.
Then the login request is sent encrypted.

The server will reply with unecrypted messages because during encryption level negotiation, the client does not ask for full encryption.

Signed-off-by: Thomas Segismont <tsegismont@gmail.com>