Implement encryption level negotiation
When the server replies to the prelogin message, it gives the desired encryption level.

If the client is configured to use encryption and the server does not support it, we must terminate the connection.
Otherwise, we enable encryption and send the login the message.

Immediately after sending the login message, we must disable encryption if it is not required for the connection lifetime.
In this case the server will send the login response unencrypted.

Signed-off-by: Thomas Segismont <tsegismont@gmail.com>