MSSQL failure when using streaming on non cursor-friendly statements
Fixes #1220

When using cursors or streaming, the client asks the server to create a cursor for results.
However, in some cases, the server cannot create it and, instead of failing, returns a complete result set at once.

In this PR, the INFO token is parsed to and the error number inspected to determine if the server executed SQL directly.
In this case, the cursor fetch request is skipped and the command is completed immediately after parsing all tokens.

Signed-off-by: Thomas Segismont <tsegismont@gmail.com>