Introduce Accumulator data structure
It is like a collection but append-only.

In order to compare performances, we have two implementations:

- one backed by an ArrayList (same behavior as before)
- one which accumulates items in chunks (avoids growing an array list and copying items)

The second one can be configured with a chunk extension policy.

In the benchmarks, we compare three cases:

- array list
- chunks of fixed size
- chunks of growing size

Results show that chunked implementation is:

- significantly faster and economical regarding memory for accumulation
- slighlty slower for traversal but, unlike array list, allocates virtually nothing

The growing size extension policy does not bring much unless the
database returns many thousands of rows in a single query.

Note than when the database returns between 2 to 10 rows (which happens often),
the chunked implementation allocates slightly more (because it needs an
extra field in the array for the next array pointer).
However, in this particular case, it is significantly faster than array
list for both accumulation and traversal.

Signed-off-by: Thomas Segismont <tsegismont@gmail.com>