Close one-shot factories when connection is closed
Fixes #1302

A close future is created for connection factories and another one for connections.
In one-shot connection factories, the factory close future is added as a hook to the connection close future.
And then the factory close future is added as context close hook, instead of the factory itself.

The combination of these changes guarantees that:

- the factory is always closed properly (and thus the underlying net client)
- factory objects aren't accumulated as close hooks in context

Signed-off-by: Thomas Segismont <tsegismont@gmail.com>