SqlConnectOptions do extend NetClientOptions which eventually leads to use a TCP client per socket forcing to keep a map of clients in the connection factory. We should avoiding this design and instead move the relevant bits of NetClientOptions to SqlConnectOptions like the new ClientSSLOptions.
After this, SqlConnectOptions does not extend NetClientOptions anymore and we can use a single TCP client in the connection factory which makes this easier to use. Thanks to the new TCP ConnectOptions the connection factory can specify the ClientSSLOptions when upgrading the connection to SSL.

We will provide another way to specify NetClientOptions for a pool as this can still be useful for other reasons.