Pg Client: fail cleanly when number of params exceeds limit
See 1142

When the number of params exceeds the limit, Postgres successfully prepares the query but returns a number of expected equal to (actual - max - 1).

Then the client reports a strange message like:

"The number of parameters to execute should be consistent with the expected number of parameters = [64464] but the actual number is [130000]."

With this change, an explicit error message is reported.

Signed-off-by: Thomas Segismont <tsegismont@gmail.com>