Fix mssql containers setup
Failures could be seen both on the developer machine and on CI.

Microsoft has changed the path to the sqlcmd utility in the latest releases of the image.

Besides, they have made connecting on an encrypted connection with sqlcmd mandatory by default.
Therefore, we must configure the utility to make this optional (-No).

Also, when now certificate are always validated but test servers use self-signed certificates.
For this reason, the utility must not try to validate the server certificate (-C).

Lastly, unlike the mcr.microsoft.com/mssql/server image, the mcr.microsoft.com/mssql-tools image has not been updated.
So the latest image release still uses the old sqlcmd utility path and options.

To avoid discrepancies in our setup, we'll use the mcr.microsoft.com/mssql/server image everywhere from now on.
We don't care too much about the size, because we need this image to run the server anyway.

Signed-off-by: Thomas Segismont <tsegismont@gmail.com>