Improve testCloseOnUndeploy
Sometimes we can see a stack trace in build logs.

It happens because the test completes before the verticle is fully undeployed.
So when the tearDown method closes vertx, another attempt to deploy the verticle is made, which prints a failure to the console, because the original verticle undeployment has been completed concurrently.

Signed-off-by: Thomas Segismont <tsegismont@gmail.com>