simple fix of oracle stream API for RAW column (#1498)
* simple fix of oracle stream API for RAW column

Signed-off-by: turakamou <turakamou@gmail.com>

* fixup! simple fix of oracle stream API for RAW column

Oracle will return "[B" as class name for byte[], I don't know why the
class loader is not able to load this.
So when we try to call getObject for this column, it is called with null
class argument.
So let's return the correct class in this case.

Signed-off-by: turakamou <turakamou@gmail.com>

* fixup! simple fix of oracle stream API for RAW column

Fix unit test for master

Signed-off-by: turakamou <turakamou@gmail.com>

* simple fix of oracle stream API for RAW column

Take into account tsegismont review comments

Signed-off-by: turakamou <turakamou@gmail.com>

---------

Signed-off-by: turakamou <turakamou@gmail.com>