Let the vertx-pg-client encoder estimate the capacity of buffer it allocates.
Motivation:

The pg encoder allocates Netty buffer without an initial capacity, when pipelining is used this buffer can be reallocated multiple times and in agressive scenario can lead to out of memory errors with the adapative allocator.

Changes:

Cumulate and encode all the outbound messages at once when flush happens instead of cumulating them in the outbound buffer.

Capacity is estimated when a message is enqueued.

At flush time, a single buffer of the estimated capacity is created and the outbound messages are written to this buffer.

This requires to pre-render some messages to estimate their length, e.g. a json object is pre-rendered to a string that can be then estimated. The client has been modified to handle message preparation which takes care of this and lazy duplicate the tuple / list of tuples when this happens.