Simplify and optimize the command response future/promise flow.
Motivation:

The command response flow uses a futures that piggies back to the query result builder and then completes the overall result promise.

Only the terminal node needs to be a promise/future as we are always on the same connection context.

The internal command response flow can uses instead a Completable flow that is much more straightforward for debugging and saves un-necessary method calls.

Changes:

Use a Completable response flow internally instead of the Promise based flow.