Decouple ConnectionFactory from connection API.
Motivation:

The ConnectionFactory SPI currently provides an SqlConnection, most callers actually care about the actual Connection SPI and unwrap it then cast it.

Instead we can have the ConnectionFactory only care about providing Connection and let callers wrap it with an SqlConnection of their choice.

As consequence this simplifies the pool implementation as they operation on Connection.