Refactor socket based clients to produce a command message from a command.
Motivation:

Commands are currently sent to every socket based encoder (e.g. PgEncoder) that then turns it into a codec object that is then encoded.

There are two issues related to this:

1. the codec is actually the message to be sent and not the command
2. the command requires the handler to be executed and we want to make the command base handler private (so it can be removed later)

Then:

Most likely we will continue this refactor by renaming *Codec classes to *Message and make the Completable<R> handler part of the message, so we can remove it from CommandBase