Make RowBase and RowDesc usable without creating subclasses.
Motivation:

RowBase is abstract and does not implement a couple of methods from its contract, such implementation actually relies on a description field present in each of its subclasses. Instead we can have RowBase carry the description field and implement those methods, as consequence RowBase can be reused as such and does not need to be abstract.

RowDesc is simply abstract but is usable out of the box, just make it not abstract so it can be instantiated right away.