Let the socket connection take care of the lifecycle of command message and the associated flow.
Motivation:

Currently each codec takes care of failing the inflight command messages when a failure occurs in the pipeline. The socket connection already takes care of translating command responses to completion of the command handler.

We should let the socket connection take care of handling the case of failure, this avoids propagating the command response handler to each codec and the codec only cares about writing an inbound command response message.