MySQL cursor fetch uses wrong row descriptor after initial fetch (#1562)
See #1525

MySQL sends column definitions in response to a prepare command.
But when executing the statement or fetching the first page of a cursor, it sends column definitions again.

The row decoder should not use the definitions given in response to prepare command because they may not be accurate.

The previous cursor implementation used them when fetching the second page.

Signed-off-by: Thomas Segismont <tsegismont@gmail.com>