Fix MetricsTestBase
Begin/End metrics size depends on the timeout parameter value.

When a timeout is expected, a single begin/end metric should be listed, corresponding to the getConnection/close calls wrapping the tasks.

When a timeout isn't expected, the number of begin/end metric is the number of tasks plus one, corresponding to the getConnection/close calls.

Signed-off-by: Thomas Segismont <tsegismont@gmail.com>