Sql Server Client may report an IndexOutOfBoundsException when using cursor (#1583)
See #1582

When using cursors, and the number of columns in the query result is a multiple of 8, and some of them have the NULL value, the client may report an IndexOutOfBoundsException.

This happens because with cursors, there is an extra ROWSTAT column on the wire that wasn't accounted for by the NBCROW decoder.

Signed-off-by: Thomas Segismont <tsegismont@gmail.com>